requirements.txt
ckantoolkit==0.0.3
